+++ packages/ako_optional/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!--
 ### Added
+* `tryOptionalOrElse` helper to resolve a optional future, catching exceptions
+* `tryOrElse` helper to resolve a future, catching exceptions
 ### Changed
 ### Deprecated
 ### Removed